+++ test/element/bootstrap3_test.rb
@@ -143,7 +143,7 @@ describe Formular::Element::Bootstrap3 do
 
     it '#error should return the error element for :body' do
       element = builder.error(:body)
-      element.to_s.must_equal %(<span class="help-block">This really isn't good enough!</span>)
+      element.to_s.must_equal %(<span class="help-block">This really isn&#39;t good enough!</span>)
     end
   end