+++ test/element/bootstrap4_test.rb
@@ -77,7 +77,7 @@ describe Formular::Element::Bootstrap4 do
 
     it '#error should return the error element for :body' do
       element = builder.error(:body)
-      element.to_s.must_equal %(<span class="form-control-feedback">This really isn't good enough!</span>)
+      element.to_s.must_equal %(<span class="form-control-feedback">This really isn&#39;t good enough!</span>)
     end
   end