+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ### Fixed
 
+* Select elements prompt and include_blank options now operate the same way as simple form
 * added procces_option to the Element API
+* Escape html chars in error_text, hint_text, input values
+
 ### Internal
 * element#options now includes both elements and html attributes
 * elements#attributes only includes html attributes and can't be mutated. Change element#options instead
 * elements#normalize_options no longer tries to use the default value if an option is present
+* added a module for html_escape
 
 # v0.2.1 2016-09-29
 
@@ -14,7 +18,6 @@
 
 * Correctly require declarative heritage for Formular::Element::Module - (@fran-worley)
 * Update readme & include example gemgem links - (@fran-worley)
-
 
 # v0.2.0 2016-09-27